python/langbot/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
google-cloud-storage
python-telegram-bot